# [optimize](util) Faster bit_pack