--- a/be/src/util/frame_of_reference_coding.h
+++ b/be/src/util/frame_of_reference_coding.h
@@ -99,6 +99,9 @@ public:
 
     void put_batch(const T* value, size_t count);
 
+    // for test
+    void test_bit_pack(const T* input, uint8_t in_num, int bit_width, uint8_t* output);
+
     // Flushes any pending values to the underlying buffer.
     // Returns the total number of bytes written
     uint32_t flush();
@@ -116,6 +119,14 @@ public:
 
 private:
     void bit_pack(const T* input, uint8_t in_num, int bit_width, uint8_t* output);
+
+    void bit_pack_8(const T* input, uint8_t in_num, int bit_width, uint8_t* output);
+
+    void bit_pack_16(const T* input, uint8_t in_num, int bit_width, uint8_t* output);
+
+    void bit_pack_32(const T* input, uint8_t in_num, int bit_width, uint8_t* output);
+
+    void bit_pack_128(const T* input, uint8_t in_num, int bit_width, uint8_t* output);
 
     void bit_packing_one_frame_value(const T* input);