… be/CMakeLists.txt
… be/benchmark/benchmark_main.cpp
… be/test/util/frame_of_reference_coding_test.cpp