@hello-stephen 2025-07-09
# BE UT Coverage Report
Increment line coverage `100.00% (102/102)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/2c517bf26945abd5a3f255916a9215a78174675b_2c517bf26945abd5a3f255916a9215a78174675b/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/2c517bf26945abd5a3f255916a9215a78174675b_2c517bf26945abd5a3f255916a9215a78174675b/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 57.34% (15513/27054) |
| Line Coverage     | 46.35% (141099/304436) |
| Region Coverage   | 45.60% (71327/156426) |
| Branch Coverage   | 40.33% (37600/93220) |